Updated sterereviwe naar double

+++ database/migrations/2025_02_28_131310_create_database_fields.php
@@ -73,7 +73,7 @@ return new class extends Migration
             $table->foreign('AccessoireId')
                 ->references('AccessoireId')->on('accessoires')
                 ->onDelete('cascade');
-            $table->integer('SterrenAantal');
+            $table->double('SterrenAantal');
             $table->date('Datum');
             $table->text('Beschrijving');
             $table->timestamps();
